backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0,<1.0.0
requests>=2.31.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
pyyaml>=6.0.1,<7.0.0
httpx>=0.27.0,<1.0.0
flexpolyline>=1.0.1,<2.0.0
